DTO para cliente removido, atualizado Id do user de Long para Integer

+++ api/src/main/java/br/ufpr/api/dto/ChamadoCreateUpdateDTO.java
@@ -7,8 +7,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record ChamadoCreateUpdateDTO(
-    @NotNull Long clienteId,
-    Long funcionarioId,                 // opcional
+    @NotNull Integer clienteId,
+    Integer funcionarioId,                 // opcional
     @NotNull Integer categoriaId,
     @NotBlank String descricaoEquipamento,
     @NotBlank String descricaoFalha,
